+++ Casks/e/easylaunchpad.rb
@@ -1,6 +1,6 @@
 cask "easylaunchpad" do
-  version "2.0.0"
-  sha256 "b7485e23738ac325f2fbe328e6d4ac22067d2793b7996c7c76d11dbddc4776c9"
+  version "2.1.0"
+  sha256 "080f9b9a2d13ee48082a2e6bf08ee3b41b72fa6d2a3b55eb35ed563e5f9539c1"
 
   url "https://github.com/NonchalantLudens/EasyLaunchPad/releases/download/v#{version}/EasyLaunchPad-#{version}.dmg"
   name "EasyLaunchPad"